refactor: Move precompute to utils package
