--- alvoc/core/precompute/__init__.py
@@ -1,1 +1,0 @@
-from alvoc.core.precompute.precompute import precompute
