+++ alvoc/core/utils/precompute.py
